entire signed signature matching
Seed matched!
Master_pubkey matched!
Master_chaincode matched!
M44_pubkey matched!
M441_pubkey matched!
M4410_pubkey matched!
M44100_pubkey matched!
M441000_pubkey matched!

public key[33 bytes]: 02b97a7f40dfd0a9989143797ded1ba7abc9105f5fc8b87ac2fce695de29684902

private key[32 bytes]: 59809ff18c189ef9a6f15724c81d0eed4f6f2352e3e92ea8dd69d0ccb87e2283

pubkeyHash[33 bytes]: 243ca4781babae0c814048a296945b9228d5104c00000000000000000000000040

unsigned txn[148 bytes] : 0200000001223ebf37da5987ed45ec2bdee33697e6fdd752823b645d545cac8994ff158c88110000001976a914d96ad3c56a2d03446c0192712119b6741d3d9ee788acffffffff0260ea0000000000001976a914ed614881f32c024a80d1b6b58dfed8f493f41c7288ac95a14200000000001976a91499ccf9022fe5173d2194659687382f235169bc5788ac0000000001000000
Unsign_txn hash1 matched!
Unsign_txn hash2 matched!

unsigned txn double hashed[32 bytes]: 2fcd6fb2835518528d7d1f396c2f61899be99f829eff0b0900c3e5702cb91391

signature raw[66 bytes]: ad14a660a926b92bbe8ced3350412d35dffa57db1cb3ea7a7df5f0a479fcdf1a117cdebba30f1db7eaa9a6978b05a59535ec757ba350149d3322dbbcac0c26af802b
Transaction signing successful.

signature[71 bytes]: 3045022100ad14a660a926b92bbe8ced3350412d35dffa57db1cb3ea7a7df5f0a479fcdf1a0220117cdebba30f1db7eaa9a6978b05a59535ec757ba350149d3322dbbcac0c26af

scriptSig[107 bytes]: 483045022100ad14a660a926b92bbe8ced3350412d35dffa57db1cb3ea7a7df5f0a479fcdf1a0220117cdebba30f1db7eaa9a6978b05a59535ec757ba350149d3322dbbcac0c26af012102b97a7f40dfd0a9989143797ded1ba7abc9105f5fc8b87ac2fce695de29684902

unsigned txn[148 bytes]: 0200000001223ebf37da5987ed45ec2bdee33697e6fdd752823b645d545cac8994ff158c88110000001976a914d96ad3c56a2d03446c0192712119b6741d3d9ee788acffffffff0260ea0000000000001976a914ed614881f32c024a80d1b6b58dfed8f493f41c7288ac95a14200000000001976a91499ccf9022fe5173d2194659687382f235169bc5788ac0000000001000000

new scriptSig[107 bytes]: 483045022100ad14a660a926b92bbe8ced3350412d35dffa57db1cb3ea7a7df5f0a479fcdf1a0220117cdebba30f1db7eaa9a6978b05a59535ec757ba350149d3322dbbcac0c26af012102b97a7f40dfd0a9989143797ded1ba7abc9105f5fc8b87ac2fce695de29684902

signed txn[230 bytes]: 0200000001223ebf37da5987ed45ec2bdee33697e6fdd752823b645d545cac8994ff158c88110000006b483045022100ad14a660a926b92bbe8ced3350412d35dffa57db1cb3ea7a7df5f0a479fcdf1a0220117cdebba30f1db7eaa9a6978b05a59535ec757ba350149d3322dbbcac0c26af012102b97a7f40dfd0a9989143797ded1ba7abc9105f5fc8b87ac2fce695de29684902ffffffff0260ea0000000000001976a914ed614881f32c024a80d1b6b58dfed8f493f41c7288ac95a14200000000001976a91499ccf9022fe5173d2194659687382f235169bc5788ac0000000001000000

verify txn[226 bytes]: 0200000001223ebf37da5987ed45ec2bdee33697e6fdd752823b645d545cac8994ff158c88110000006b483045022100ad14a660a926b92bbe8ced3350412d35dffa57db1cb3ea7a7df5f0a479fcdf1a0220117cdebba30f1db7eaa9a6978b05a59535ec757ba350149d3322dbbcac0c26af012102b97a7f40dfd0a9989143797ded1ba7abc9105f5fc8b87ac2fce695de29684902ffffffff0260ea0000000000001976a914ed614881f32c024a80d1b6b58dfed8f493f41c7288ac95a14200000000001976a91499ccf9022fe5173d2194659687382f235169bc5788ac0000000001000000

sign matched!

+++ include/utility.h
@@ -16,7 +16,7 @@ void node_details(HDNode node);
 
 void hash256(const uint8_t *data, uint8_t *output, size_t size);
 
-void generate_scriptSig(const uint8_t *signature, uint8_t *scriptSig, uint8_t* publicKey, uint8_t sig_len, uint8_t scriptSig_len, uint8_t pubkey_len);
+void generate_scriptSig(const uint8_t *signature, uint8_t *scriptSig, uint8_t* publicKey, size_t sig_len, size_t scriptSig_len, size_t pubkey_len);
 void generate_scriptPubKey(const uint8_t *scriptSig, size_t scriptSigLen, uint8_t *scriptPubKey, uint8_t scriptPubKey_len);
 
 void concatenate_arrays(uint8_t *dest, const uint8_t *src1, size_t len1, const uint8_t *src2, size_t len2);
